+++ app/src/components/AlertMessage.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const AlertMessage = ({ children }) => (
+  <div className='new-status-message'>{children}</div>
+)
+
+export default AlertMessage