+++ app/src/pages/Detailed/DetailedGQL.js
@@ -54,10 +54,16 @@ export const projectBySlugGQL = gql`
         limit: 10
         transactionType: OUT
       ) {
-        fromAddress
+        fromAddress {
+          address
+          isExchange
+        }
+        toAddress {
+          address
+          isExchange
+        }
         trxValue
         trxHash
-        toAddress
         datetime
       }
       ethSpentOverTime(from: $fromOverTime, to: $to, interval: $interval) {
@@ -140,7 +146,7 @@ export const HistoryPriceGQL = gql`
 
 export const GithubActivityGQL = gql`
   query queryGithubActivity(
-    $ticker: String
+    $slug: String
     $from: DateTime
     $to: DateTime
     $interval: String
@@ -148,7 +154,7 @@ export const GithubActivityGQL = gql`
     $movingAverageIntervalBase: String
   ) {
     githubActivity(
-      ticker: $ticker
+      slug: $slug
       from: $from
       to: $to
       interval: $interval
@@ -192,11 +198,10 @@ export const TransactionVolumeGQL = gql`
 `
 
 export const ExchangeFundFlowGQL = gql`
-  query exchangeFundFlowGQL($slug: String, $from: DateTime, $to: DateTime) {
-    exchangeFundFlow(slug: $slug, from: $from, to: $to, transactionType: ALL) {
+  query exchangeFundsFlowGQL($slug: String, $from: DateTime, $to: DateTime) {
+    exchangeFundsFlow(slug: $slug, from: $from, to: $to) {
       datetime
-      transactionVolume
-      address
+      fundsFlow
       __typename
     }
   }