+++ test/sanbase_web/graphql/apikey_resolver_test.exs
@@ -1,5 +1,7 @@
 defmodule SanbaseWeb.Graphql.ApikeyResolverTest do
   use SanbaseWeb.ConnCase, async: false
+  @moduletag checkout_repo: [Sanbase.Repo, Sanbase.TimescaleRepo]
+  @moduletag timescaledb: true
 
   import ExUnit.CaptureLog
   import SanbaseWeb.Graphql.TestHelpers
@@ -350,31 +352,23 @@ defmodule SanbaseWeb.Graphql.ApikeyResolverTest do
   end
 
   defp init_databases(slug, datetime1, datetime2) do
-    alias Sanbase.Influxdb.Measurement
-    alias Sanbase.Etherbi.BurnRate.Store
-    alias Sanbase.Repo
-    alias Sanbase.Model.Project
+    import Sanbase.TimescaleFactory
 
-    Store.create_db()
+    contract_address = "0" <> Sanbase.TestUtils.random_string()
 
-    contract_address = "0x123123123"
+    insert(:burn_rate, %{
+      contract_address: contract_address,
+      timestamp: datetime1,
+      burn_rate: 5000
+    })
 
-    Store.import([
-      %Measurement{
-        timestamp: datetime1 |> DateTime.to_unix(:nanoseconds),
-        fields: %{burn_rate: 5000},
-        tags: [],
-        name: contract_address
-      },
-      %Measurement{
-        timestamp: datetime2 |> DateTime.to_unix(:nanoseconds),
-        fields: %{burn_rate: 1000},
-        tags: [],
-        name: contract_address
-      }
-    ])
+    insert(:burn_rate, %{
+      contract_address: contract_address,
+      timestamp: datetime2,
+      burn_rate: 1000
+    })
 
-    %Project{
+    %Sanbase.Model.Project{
       name: "Santiment",
       ticker: "SAN",
       coinmarketcap_id: slug,